docs: listed below. - author information and git repo information added in package.json. - `jsdoc2md` config updated. - `DOCUMENTATION` file separated. - jsDoc updated for `blockBrowserBackButton`, `createTable`, `getNestedValue`, `getUniqueArray`, `isURLValid`, `mergeObjects`, `objectToQueryParams`, `printPretty`, `removeEmptyProperty`, `reverseStr`, `shortenString`, `shuffleString`, `sortObjectByKey`, `uniqueArrayOfObjects`.
