+++ docs/template.hbs
@@ -1,37 +1,2 @@
-# ez.js
-
-<div id="badges">
- <img alt="npm" src="https://img.shields.io/npm/dm/@bhairesh/ez.js">
-</div>
-
-The `ez.js` library exported as cjs, esm and umd modules.
-
-## Installation
-
-Using npm:
-```shell
-npm i -g @bhairesh/ez.js
-npm i @bhairesh/ez.js
-```
-
-Imports:
-```js
-// Load the package.
-const ezjs = require("@bhairesh/ez.js");
-
-// or
-
-import ezjs from "@bhairesh/ez.js";
-```
-
-## Why ez.js?
-
-`ez.js` makes JavaScript easier by taking the hassle out of working with arrays, numbers, objects, strings, etc.
-
-## Links
-
-[Contribution](./docs/CONTRIBUTION.md)
-
-## Documentation
 
 {{>main}}
