+++ jsdoc2md.json
@@ -4,7 +4,7 @@
     "cleverLinks": false,
     "monospaceLinks": false
   },
-  "recurseDepth": 10,
+  "recurseDepth": 0,
   "tags": {
     "allowUnknownTags": true,
     "dictionaries": ["jsdoc"]
@@ -15,6 +15,6 @@
   "opts": {
     "encoding": "utf8",
     "destination": "docs/",
-    "recurse": true
+    "recurse": false
   }
 }
