+++ src/helpers/getUniqueArray.js
@@ -1,10 +1,13 @@
 /**
  * @param {Array} arr
+ * @returns {Array}
  */
-module.exports = function getUniqueArray(arr = []) {
+function getUniqueArray(arr = []) {
   const uArr = [];
   arr.forEach((a) => {
     if (uArr.indexOf(a) === -1) uArr.push(a);
   });
   return uArr;
-};
+}
+
+module.exports = getUniqueArray;
