+++ src/helpers/index.js
@@ -1,26 +1,29 @@
-export { default as getNestedValue } from "./getNestedValue.js";
-export { default as dataFormatter } from "./dataFormatter.js";
-export { default as currencyFormatter } from "./currencyFormatter.js";
-export { default as uniqueArrayOfObjects } from "./uniqueArrayOfObjects.js";
-export { default as compareObjects } from "./compareObjects.js";
+export { default as add2Matrix } from "./add2Matrix.js";
+export { default as addDelEleFromArray } from "./addDelEleFromArray.js";
 export { default as arrayIntoChunks } from "./arrayIntoChunks.js";
+export { default as blockBrowserBackButton } from "./blockBrowserBackButton.js";
+export { default as camelCase } from "./camelCase.js";
+export { default as compareObjects } from "./compareObjects.js";
+export { strToNum, toNumber } from "./convert.js";
+export { default as createTable } from "./createTable.js";
+export { default as currencyFormatter } from "./currencyFormatter.js";
+export { default as dataFormatter } from "./dataFormatter.js";
+export { default as dateDiff } from "./dateDiff.js";
+export { default as generateRandomString } from "./generateRandomString.js";
+export { default as getDiffInHrs } from "./getDiffInHrs.js";
+export { default as getNestedValue } from "./getNestedValue.js";
+export { default as getUniqueArray } from "./getUniqueArray.js";
+export { default as hasOwnProperty } from "./hasOwnProperty.js";
+export { default as isEmpty } from "./isEmpty.js";
+export { default as isStr } from "./isStr.js";
+export { default as isStrHasSpecialChar } from "./isStrHasSpecialChar.js";
+export { default as isURLValid } from "./isURLValid.js";
+export { default as mergeObjects } from "./mergeObjects.js";
+export { default as objectToQueryParams } from "./objectToQueryParams.js";
+export { default as printPretty } from "./printPretty.js";
+export { default as removeEmptyProperty } from "./removeEmptyProperty.js";
+export { reverseEachWordInSentence, reverseSentence, reverseStr } from "./reverseStr.js";
+export { default as shortenString } from "./shortenString.js";
 export { default as sortObjectByKey } from "./sortObjectByKey.js";
 export { default as sortObjectByMultipleKeys } from "./sortObjectByMultipleKeys.js";
-export { default as createTable } from "./createTable.js";
-export { default as removeEmptyProperty } from "./removeEmptyProperty.js";
-export { strToNum, toNumber } from "./convert.js";
-export { default as shuffleString } from "./shuffleString.js";
-export { default as isStrHasSpecialChar } from "./isStrHasSpecialChar.js";
-export { default as hasOwnProperty } from "./hasOwnProperty.js";
-export { default as objectToQueryParams } from "./objectToQueryParams.js";
-export { default as generateRandomString } from "./generateRandomString.js";
-export { default as shortenString } from "./shortenString.js";
-export { default as isURLValid } from "./isURLValid.js";
-export { default as camelCase } from "./camelCase.js";
-export { default as getDateDifferenceInHours } from "./getDiffInHrs.js";
-export { default as getDateDifference } from "./dateDiff.js";
-export { default as addDelEleFromArray } from "./addDelEleFromArray.js";
-export { reverseStr, reverseEachWordInSentence, reverseSentence } from "./reverseStr.js";
-export { default as isEmpty } from "./isEmpty.js";
-export { default as blockBrowserBackButton } from "./blockBrowserBackButton.js";
-export { default as printPretty } from "./printPretty.js";
+export { default as uniqueArrayOfObjects } from "./uniqueArrayOfObjects.js";
