+++ src/helpers/mergeObjects.js
