+++ src/helpers/objectToQueryParams.js
@@ -1,9 +1,12 @@
 /**
  * @param {object} o
+ *
+ * @returns {string}
  */
-const objectToQueryParams = (o = {}) =>
-  Object.entries(o)
+function objectToQueryParams(o = {}) {
+  return Object.entries(o)
     .map((p) => `${encodeURIComponent(p[0])}=${encodeURIComponent(p[1])}`)
     .join("&");
+}
 
 module.exports = objectToQueryParams;
