+++ src/helpers/reverseStr.js
@@ -1,5 +1,6 @@
 /**
  * @param {string} str: string without space
+ * @returns {string}
  */
 function reverseStr(str) {
   return str.split("").reverse().join("");
@@ -7,6 +8,7 @@ function reverseStr(str) {
 
 /**
  * @param {string} sentence: pass a sentence.
+ * @returns {string}
  */
 function reverseEachWordInSentence(sentence) {
   return sentence.split(" ").map(reverseStr).join(" ");
@@ -14,6 +16,7 @@ function reverseEachWordInSentence(sentence) {
 
 /**
  * @param {string} sentence - pass a sentence.
+ * @returns {string}
  */
 function reverseSentence(sentence) {
   return sentence.split(" ").reverse().join(" ");
