+++ src/helpers/shortenString.js
@@ -1,6 +1,7 @@
 /**
  * @param {string} str
  * @param {number} len - default is 50
+ * @returns {string}
  */
 const shortenString = (str = "", len = 50) => {
   if (str.length > len) return str.substring(0, len);
